(*) done restructuring sodtube
