+++ app/drivers/hydro/main_driver.cc
@@ -37,6 +37,7 @@
 #include "flecsi/data/data_client.h"
 #include "flecsi/data/data.h"
 
+// #define poly_gamma 5./3.
 #include "params.h"
 #include "bodies_system.h"
 #include "default_physics.h"
@@ -64,7 +65,6 @@ void set_derived_params() {
   physics::iteration = initial_iteration;
   physics::totaltime = initial_time;
   physics::dt = initial_dt; // TODO: use particle separation and Courant factor
-
 }
 
 namespace flecsi{
@@ -91,74 +91,86 @@ mpi_init_task(const char * parameter_file){
   body_system<double,gdimension> bs;
   bs.read_bodies(initial_data_file.c_str(),initial_iteration);
 
-  // Set the boundaries to be at 10% of the total range
-  double h = bs.getSmoothinglength();
-  auto range_boundaries = bs.getRange();
-  double distance = fabs(range_boundaries[1][0]-range_boundaries[0][0]);
-  physics::min_boundary = {range_boundaries[0][0]+distance*0.1+2*h};
-  physics::max_boundary = {range_boundaries[1][0]-distance*0.1-2*h};
-
-  clog_one(info) << "Boundaries=" << physics::min_boundary << ";"
-         << physics::max_boundary << std::endl;
-
-  bs.update_iteration();
-
 #ifdef OUTPUT
   bs.write_bodies(output_h5data_prefix,physics::iteration);
 #endif
 
   ++physics::iteration;
-  do
-  {
+  do {
     analysis::screen_output();
     MPI_Barrier(MPI_COMM_WORLD);
 
+    if (physics::iteration == 1){
+      // Compute and prepare the tree for this iteration
+      // - Compute the Max smoothing length
+      // - Compute the range of the system using the smoothinglength
+      // - Compute the keys
+      // - Distributed qsort and sharing
+      // - Generate and feed the tree
+      // - Exchange branches for smoothing length
+      // - Compute and exchange ghosts in real smoothing length
+      bs.update_iteration();
 
-    // Compute and prepare the tree for this iteration
-    // - Compute the Max smoothing length
-    // - Compute the range of the system using the smoothinglength
-    // - Compute the keys
-    // - Distributed qsort and sharing
-    // - Generate and feed the tree
-    // - Exchange branches for smoothing length
-    // - Compute and exchange ghosts in real smoothing length
-    bs.update_iteration();
-
-    clog_one(trace) << "compute_density_pressure_soundspeed" << std::flush;
-    bs.apply_in_smoothinglength(physics::compute_density_pressure_soundspeed);
-    clog_one(trace) << ".done" << std::endl;
-
-    bs.update_neighbors();
-
-    clog_one(trace) << "Hydro acceleration" << std::flush;
-    bs.apply_in_smoothinglength(physics::compute_hydro_acceleration);
-    clog_one(trace) << ".done" << std::endl;
-
-    clog_one(trace) << "Internalenergy" << std::flush;
-    bs.apply_in_smoothinglength(physics::compute_dudt);
-    clog_one(trace) << ".done" << std::endl;
-
-    if(physics::iteration == initial_iteration + 1){
-      clog_one(trace) << "leapfrog" << std::flush;
-      bs.apply_all(physics::leapfrog_integration_first_step);
+      // at the initial iteration, P, rho and cs have not been computed yet;
+      // for all subsequent steps, however, they are computed at the end 
+      // of the iteration
+      clog_one(trace) << "first iteration: pressure, rho and cs" << std::flush;
+      bs.apply_in_smoothinglength(physics::compute_density_pressure_soundspeed);
+      bs.apply_all(physics::save_velocityhalf);
       clog_one(trace) << ".done" << std::endl;
-    }else{
-      clog_one(trace) << "leapfrog" << std::flush;
-      bs.apply_all(physics::leapfrog_integration);
+
+      // necessary for computing dv/dt and du/dt in the next step
+      bs.update_neighbors();
+
+      clog_one(trace) << "compute accelerations and dudt" << std::flush;
+      bs.apply_in_smoothinglength(physics::compute_hydro_acceleration);
+      bs.apply_in_smoothinglength(physics::compute_dudt);
+      clog_one(trace) << ".done" << std::endl;
+
+    }
+    else {
+      clog_one(trace) << "leapfrog: kick one" << std::flush;
+      bs.apply_all(physics::leapfrog_kick_v);
+      bs.apply_all(physics::leapfrog_kick_u);
+      bs.apply_all(physics::save_velocityhalf);
+      clog_one(trace) << ".done" << std::endl;
+
+      // sync velocities
+      bs.update_neighbors();
+
+      clog_one(trace) << "leapfrog: drift" << std::flush;
+      bs.apply_all(physics::leapfrog_drift);
+      bs.apply_in_smoothinglength(physics::compute_density_pressure_soundspeed);
+      clog_one(trace) << ".done" << std::endl;
+
+      // recompute the tree and sync
+      bs.update_iteration();
+      bs.update_neighbors();
+
+      clog_one(trace) << "leapfrog: kick two (velocity)" << std::flush;
+      bs.apply_in_smoothinglength(physics::compute_hydro_acceleration);
+      bs.apply_all(physics::leapfrog_kick_v);
+      clog_one(trace) << ".done" << std::endl;
+
+      // sync velocities
+      bs.update_neighbors();
+
+      clog_one(trace) << "leapfrog: kick two (int. energy)" << std::flush;
+      bs.apply_in_smoothinglength(physics::compute_dudt);
+      bs.apply_all(physics::leapfrog_kick_u);
       clog_one(trace) << ".done" << std::endl;
     }
-
-    clog_one(trace) << "dudt integration" << std::flush;
-    bs.apply_all(physics::dudt_integration);
-    clog_one(trace) << ".done" << std::endl;
 
 #ifdef OUTPUT_ANALYSIS
     // Compute the analysis values based on physics
     bs.get_all(analysis::compute_lin_momentum);
     bs.get_all(analysis::compute_total_mass);
+    bs.get_all(analysis::compute_total_energy);
+    bs.get_all(analysis::compute_total_ang_mom);
     // Only add the header in the first iteration
-    analysis::scalar_output("scalar_reductions.dat"); // TODO: hardcoded value
+    analysis::scalar_output("scalar_reductions.dat");
 #endif
+
 
 #ifdef OUTPUT
     if(out_h5data_every > 0 && physics::iteration % out_h5data_every == 0){
@@ -170,14 +182,14 @@ mpi_init_task(const char * parameter_file){
     physics::totaltime += physics::dt;
 
   } while(physics::iteration <= final_iteration);
-}
+} // mpi_init_task
+
 
 flecsi_register_mpi_task(mpi_init_task);
 
-
-void
+void 
 usage() {
-  clog_one(warn) << "Usage: ./sodtube <parameter-file.par>"
+  clog_one(warn) << "Usage: ./noh <parameter-file.par>"
                  << std::endl << std::flush;
 }
 
@@ -186,17 +198,17 @@ void
 specialization_tlt_init(int argc, char * argv[]){
   clog_one(trace) << "In user specialization_driver" << std::endl;
 
-  // check options list: exactly one command-line argument is accepted
+  // check options list: exactly one option is allowed
   if (argc != 2) {
     clog_one(error) << "ERROR: parameter file not specified!" << std::endl;
     usage();
     return;
   }
 
-  // pass const char * argv[1] to mpi_init_task
   flecsi_execute_mpi_task(mpi_init_task, argv[1]);
 
 } // specialization driver
+
 
 void
 driver(int argc,  char * argv[]){
