+++ app/drivers/include/user.h
@@ -17,35 +17,18 @@
  *~--------------------------------------------------------------------------~*/
 
 /**
- * @file main.cc
+ * @file user.h
  * @author Julien Loiseau
  * @date April 2017
- * @brief Main function, start MPI with Gasnet. Then launch fleCSI runtime.  
+ * @brief User define for dimension and type 
  */
 
-#include <flecsi.h>
-#include "flecsi/execution/execution.h"
-#include "flecsi/concurrency/thread_pool.h"
+#ifndef _user_h_
+#define _user_h_
 
-#if FLECSI_RUNTIME_MODEL == FLECSI_RUNTIME_MODEL_legion
-  #include <mpi.h>
-  #include <legion.h>
-#endif
+#define INTERNAL_ENERGY
+#define OUTPUT
+static const size_t gdimension = EXT_GDIMENSION;
+using type_t = double;
 
-int main(int argc, char * argv[]){
-  int provided;
-  // Normal way 
-  MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &provided);
-  if (provided < MPI_THREAD_MULTIPLE)
-    printf("ERROR: Your implementation of MPI does not support "
-     "MPI_THREAD_MULTIPLE which is required for use of the "
-     "GASNet MPI conduit with the Legion-MPI Interop!\n");
-  assert(provided == MPI_THREAD_MULTIPLE);
-
-  auto retval = flecsi::execution::context_t::instance().initialize(argc,argv);
-
-  return retval;
-
-}
-
-
+#endif // _user_h_
