+++ config/project.cmake
@@ -115,6 +115,6 @@ cinch_add_application_directory("app/sedov")
 cinch_add_application_directory("app/noh")
 cinch_add_application_directory("app/fluid_3D")
 cinch_add_application_directory("app/fluid_2D")
-cinch_add_application_directory("app/sodtube")
 cinch_add_application_directory("app/bwd")
 cinch_add_application_directory("app/id_generators")
+cinch_add_application_directory("app/drivers")
